+++ src/components/GeneralComponents/Footer.svelte
