+++ src/components/ShoppingCart/CloseButton.svelte
@@ -11,7 +11,7 @@
     top: -3rem;
     right: 0;
     color: white;
-    background: purple;
+    background: #689c54;
     border: 2px solid pink;
   }
 </style>
